Lazy floor and ceiling render done
Floor and ceiling tiles are draw in a vertical box just like walls, but
illusion is complete enough with sand texture. Other textures may not
fair as well though.

Previous bug was when I adjusted the height of the ground box to be
below the wall box (or where a wall box would be in a space), I was
getting the height correct (wall.top+wall.height). But I could not tell
since the game will render a box above the player head.

This would result in me trying to lower the ground more, which would
lower the box at the player below the player (so it wouldn't appear at
all), but the ground in the distance would appear too low.

So the fix is to not draw when s=0, that is at player position. Then the
ground appears at the right height at all distances.

The non-lazy way to render the floor would be to make a copy of
camera.project to project the texture as floor tiles instead of as wall
tiles.

Alternatively, I could even do more expensive row casting (which I did
in drawRows, but resulted in a) out of position floor tiles inside walls
and b) weird artifacts on screen). To avoid it seems I need to keep both
draw calls down, and do render in a single set of stripes, vertical or
horizontal, but not both.