More render changes, bugs left
Render code modified to be similar to another raycaster fps.
Modification not complete.

A debugging red dot draws in the right location and roughly when
expected.

/Seems/ that code should draw enemies right and with some optimization,
but column coords way off of expected. As the other raycaster is known
to work, some more debugging should let ours work as well.