read code -> render fixes
So I needed to express width and height as fractions of wall spacing,
not the size of the image used. Which is why my draw coords were
astronomical.

Now the enemy renders where it should, and renders in a similar fashion
to walls.

Just need to add in and tweak some code from the example I'm using to
check which draw column is closest to player, and know when to not draw.

Basically a painter's algorithm. Remember each drawn wall column, save
xoffset and dist in a table, check table whenever drawing to same
column.