Enemy sprite culling works
By keeping track of an array of rendered column distances, enemy culling
now functions correctly! See http://i.imgur.com/bgty9PF.jpg.

An array simply holds the distance of the currently rendered object. If
another column render is closer than distance saved in columns array,
the new column is drawn over the old, and the distance updated.

The 3 ray check did not work as expected and caused 'flickering', that
is would decide to render correctly 1 frame and stop for another as the
player moved about.

By removing the 3ray check 'optimization', only the check of columns
array decides when not to render enemy columns.

One small annoyance left is when an enemy is up against a wall, of
course one half of the enemy seems imbedded in the wall. While this is a
correct render, we would want an enemy to never get so close to the
wall.

Enemies do currently move slowly away from a wall, but this is not a
final solution.

TODO: Solve above annoyance by keeping enemies 1/2 width away from
walls.

TODO: Apply same sprite render to missile sprites and other sprites.