Of course culling bugs left
While culling seems to work overall, a bug occurs when viewing an enemy
to the east.

It appears to be a failure of the FOV check, and indeed the FOV check
did have an error. But after changing the FOV check and debugging it,
this edge case remains.

The bug only occurs when viewing an enemy to the east, if the player
pivots right, the enemy ceases to render completely, it disappears from
the left side of the screen where it is still within the FOV.

Neither the FOV check or distance check seem to be the issue. Then what
needs checking is the xoffset and render parameters. It might be
rendering, but in the wrong location at that specific relation of player
and enemy angles.