Yet another render fix maybe
Error where enemy to the east does not render correctly fixed, 99% sure.

Unlike the example, I was not keeping track of all directions within the
range of -pi =< angle =< pi, but rather as the full 2*pi range.

This led to weird conversion of angles in drawEnemy() that produced this
edge case. Specifically, when angle was negative, I would convert to a
positive angle. via angle = 2*PI + angle.

This along with future checks on dAngle would produce an angle greater
than PI, and xoffsets way offscreen. So the enemy rendered but did not
appear onscreen.

I removed the faulty conversion, and a dummy check for dAngles above
1.5*PI, where the issue was occurring.

Testing shows the fix works, and that the fov limit check may not
improve fps.

A better fix to remove the many angle checks would to keep the same
reference point for directional angles, that is always within +-PI,
instead of [0, 2*PI]