Converted missile render to column method
Missiles now render as enemies and walls do, that is as a series of
columns with proper z order.

Changes to accomplish this included creating a texture data class to
describe a texture as a single image or a sprite sheet.

Methods to draw enemies are now generic enough to take any entity that
has standard x, y position, alive state, and texture data.

These methods have been renamed along the lines of drawEntity to reflect
they can draw any entity, and correctly handle drawing an image or a
sprite sheet cell as described by an entities texture data.

TODO: This should allow all entities to be grouped into a single array
with one pass to draw all columns across the FOV.